Limit impact of '@putenv' during compilation tasks
Composer-compile-plugin is based on a philosophy of encapsulation/isolation -
i.e.  that a package may internally use some compilation steps, and that it is
none of the root-package's business.  Allowing compilation tasks to put things
into a truly global environment would invite weird/hard-to-identify
interactions among packages (or, god help us, it might invite people to
intentionally create those weird/hard-to-identify interactions).